fixVue3 import via postbuild script
